move some code to other part
